modal page moved to components
