+++ client/src/components/Logo.svelte
