+++ client/src/components/Logout.svelte
