+++ client/src/pages/Dashboardpage.svelte
@@ -1,9 +1,9 @@
 <script>
-import Dashboard from "../componants/Dashboard.svelte"
-import Header from "../componants/Header.svelte"
-import Logo from "../componants/logo.svelte"
-import Logout from "../componants/logout.svelte"
-import Nav from "../componants/Nav.svelte"
+import Dashboard from "../components/Dashboard.svelte"
+import Header from "../components/Header.svelte"
+import Logo from "../components/logo.svelte"
+import Logout from "../components/logout.svelte"
+import Nav from "../components/Nav.svelte"
 </script>
 
 <div class="h-screen w-screen p-1">
